+++ include/kiss3d-custom/src/resource/mesh.rs
@@ -16,10 +16,10 @@ mod error;
 /// It also contains the GPU location of those buffers.
 pub struct Mesh {
     coords: Arc<RwLock<GPUVec<Point3<f32>>>>,
-    faces: Arc<RwLock<GPUVec<Point3<u16>>>>,
+    faces: Arc<RwLock<GPUVec<Point3<u32>>>>,
     normals: Arc<RwLock<GPUVec<Vector3<f32>>>>,
     uvs: Arc<RwLock<GPUVec<Point2<f32>>>>,
-    edges: Option<Arc<RwLock<GPUVec<Point2<u16>>>>>,
+    edges: Option<Arc<RwLock<GPUVec<Point2<u32>>>>>,
 }
 
 impl Mesh {
@@ -28,7 +28,7 @@ impl Mesh {
     /// If the normals and uvs are not given, they are automatically computed.
     pub fn new(
         coords: Vec<Point3<f32>>,
-        faces: Vec<Point3<u16>>,
+        faces: Vec<Point3<u32>>,
         normals: Option<Vec<Vector3<f32>>>,
         uvs: Option<Vec<Point2<f32>>>,
         dynamic_draw: bool,
@@ -177,7 +177,7 @@ impl Mesh {
     /// Creates a new mesh. Arguments set to `None` are automatically computed.
     pub fn new_with_gpu_vectors(
         coords: Arc<RwLock<GPUVec<Point3<f32>>>>,
-        faces: Arc<RwLock<GPUVec<Point3<u16>>>>,
+        faces: Arc<RwLock<GPUVec<Point3<u32>>>>,
         normals: Arc<RwLock<GPUVec<Vector3<f32>>>>,
         uvs: Arc<RwLock<GPUVec<Point2<f32>>>>,
     ) -> Mesh {
@@ -263,7 +263,7 @@ impl Mesh {
     }
 
     /// This mesh faces.
-    pub fn faces(&self) -> &Arc<RwLock<GPUVec<Point3<u16>>>> {
+    pub fn faces(&self) -> &Arc<RwLock<GPUVec<Point3<u32>>>> {
         &self.faces
     }
 
@@ -285,7 +285,7 @@ impl Mesh {
     /// Computes normals from a set of faces.
     pub fn compute_normals_array(
         coordinates: &[Point3<f32>],
-        faces: &[Point3<u16>],
+        faces: &[Point3<u32>],
     ) -> Vec<Vector3<f32>> {
         let mut res = Vec::new();
 
@@ -297,7 +297,7 @@ impl Mesh {
     /// Computes normals from a set of faces.
     pub fn compute_normals(
         coordinates: &[Point3<f32>],
-        faces: &[Point3<u16>],
+        faces: &[Point3<u32>],
         normals: &mut Vec<Vector3<f32>>,
     ) {
         let mut divisor: Vec<f32> = iter::repeat(0f32).take(coordinates.len()).collect();
