+++ include/kiss3d-custom/src/scene/planar_scene_node.rs
@@ -222,7 +222,7 @@ impl PlanarSceneNodeData {
     ///
     /// The provided closure is called once per object.
     #[inline(always)]
-    pub fn modify_faces<F: FnMut(&mut Vec<Point3<u16>>)>(&mut self, f: &mut F) {
+    pub fn modify_faces<F: FnMut(&mut Vec<Point3<u32>>)>(&mut self, f: &mut F) {
         self.apply_to_objects_mut(&mut |o| o.modify_faces(f))
     }
 
@@ -230,7 +230,7 @@ impl PlanarSceneNodeData {
     ///
     /// The provided closure is called once per object.
     #[inline(always)]
-    pub fn read_faces<F: FnMut(&[Point3<u16>])>(&self, f: &mut F) {
+    pub fn read_faces<F: FnMut(&[Point3<u32>])>(&self, f: &mut F) {
         self.apply_to_objects(&mut |o| o.read_faces(f))
     }
 
@@ -665,8 +665,8 @@ impl PlanarSceneNode {
                     capsule_vtx.push(Point2::new(ang.cos() * r, ang.sin() * r + h / 2.0));
                     capsule_ids.push(Point3::new(
                         0,
-                        capsule_vtx.len() as u16 - 2,
-                        capsule_vtx.len() as u16 - 1,
+                        capsule_vtx.len() as u32 - 2,
+                        capsule_vtx.len() as u32 - 1,
                     ));
                 }
 
@@ -675,12 +675,12 @@ impl PlanarSceneNode {
                     capsule_vtx.push(Point2::new(ang.cos() * r, ang.sin() * r - h / 2.0));
                     capsule_ids.push(Point3::new(
                         0,
-                        capsule_vtx.len() as u16 - 2,
-                        capsule_vtx.len() as u16 - 1,
+                        capsule_vtx.len() as u32 - 2,
+                        capsule_vtx.len() as u32 - 1,
                     ));
                 }
 
-                capsule_ids.push(Point3::new(0, capsule_vtx.len() as u16 - 1, 1));
+                capsule_ids.push(Point3::new(0, capsule_vtx.len() as u32 - 1, 1));
 
                 let capsule = PlanarMesh::new(capsule_vtx, capsule_ids, None, false);
                 let mesh = Rc::new(RefCell::new(capsule));
@@ -724,7 +724,7 @@ impl PlanarSceneNode {
         let mut indices = Vec::new();
 
         for i in 1..polygon.len() - 1 {
-            indices.push(Point3::new(0, i as u16, i as u16 + 1));
+            indices.push(Point3::new(0, i as u32, i as u32 + 1));
         }
 
         let mesh = PlanarMesh::new(polygon, indices, None, false);
@@ -824,7 +824,7 @@ impl PlanarSceneNode {
     ///
     /// The provided closure is called once per object.
     #[inline(always)]
-    pub fn modify_faces<F: FnMut(&mut Vec<Point3<u16>>)>(&mut self, f: &mut F) {
+    pub fn modify_faces<F: FnMut(&mut Vec<Point3<u32>>)>(&mut self, f: &mut F) {
         self.data_mut().modify_faces(f)
     }
 
@@ -832,7 +832,7 @@ impl PlanarSceneNode {
     ///
     /// The provided closure is called once per object.
     #[inline(always)]
-    pub fn read_faces<F: FnMut(&[Point3<u16>])>(&self, f: &mut F) {
+    pub fn read_faces<F: FnMut(&[Point3<u32>])>(&self, f: &mut F) {
         self.data().read_faces(f)
     }
